experimento1/abcjobs/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv